refactor!: reorganize util package structure
- Moved various utility classes to more specific sub-packages for better organization.
- Updated import statements across multiple files to reflect new package paths.
- Removed unused `DrivingMotor` and `UpperSubstructure` enums from the utils package.

BREAKING CHANGES:
- Remove unused `DrivingMotor` and `UpperSubstructure` enums from utils.
