+++ src/main/java/frc/robot/commands/ManualDriveCommand.java
@@ -7,7 +7,7 @@ import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.Constants.DriveConstants;
 import frc.robot.Constants.OperatorConstants;
 import frc.robot.subsystems.DriveSubsystem;
-import frc.robot.util.Vector2d;
+import frc.robot.util.math.Vector2d;
 
 public abstract class ManualDriveCommand extends Command {
