+++ src/main/java/frc/robot/commands/XboxParkerManualDriveCommand.java
@@ -6,7 +6,7 @@ import edu.wpi.first.wpilibj.XboxController;
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.Constants.OperatorConstants;
 import frc.robot.subsystems.DriveSubsystem;
-import frc.robot.util.Vector2d;
+import frc.robot.util.math.Vector2d;
 
 public class XboxParkerManualDriveCommand extends ManualDriveCommand {
