+++ src/main/java/frc/robot/commands/narwhal/NarwhalManualClimberCommand.java
@@ -3,7 +3,7 @@ package frc.robot.commands.narwhal;
 import edu.wpi.first.wpilibj.XboxController;
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.subsystems.narwhal.NarwhalClimber;
-import frc.robot.util.NarwhalClimberState;
+import frc.robot.util.upper_assembly.narwhal.NarwhalClimberState;
 
 public class NarwhalManualClimberCommand extends Command {
