+++ src/main/java/frc/robot/subsystems/UpperAssembly.java
@@ -6,7 +6,7 @@ import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.wpilibj.XboxController;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.Subsystem;
-import frc.robot.util.ScoringHeight;
+import frc.robot.util.upper_assembly.ScoringHeight;
 
 /**
  * Represents the upper assembly of the robot, which may vary depending on the selected implementation.
