+++ src/main/java/frc/robot/subsystems/narwhal/NarwhalUpperAssembly.java
@@ -7,7 +7,7 @@ import edu.wpi.first.wpilibj.XboxController;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.RunCommand;
 import frc.robot.subsystems.UpperAssembly;
-import frc.robot.util.ScoringHeight;
+import frc.robot.util.upper_assembly.ScoringHeight;
 
 public class NarwhalUpperAssembly implements UpperAssembly {
