+++ src/main/java/frc/robot/subsystems/squid/SquidUpperAssembly.java
@@ -9,7 +9,7 @@ import edu.wpi.first.wpilibj2.command.ParallelCommandGroup;
 import edu.wpi.first.wpilibj2.command.RunCommand;
 import frc.robot.commands.squid.ManualSquidManipulatorCommand;
 import frc.robot.subsystems.UpperAssembly;
-import frc.robot.util.ScoringHeight;
+import frc.robot.util.upper_assembly.ScoringHeight;
 
 public class SquidUpperAssembly implements UpperAssembly {
