+++ src/main/java/frc/robot/util/LEDs/LEDConfiguration.java
@@ -1,4 +1,4 @@
-package frc.utils.LEDs;
+package frc.robot.util.LEDs;
 
 import static edu.wpi.first.units.Units.Seconds;
