+++ src/main/java/frc/robot/util/math/RateLimiter2d.java
@@ -1,4 +1,4 @@
-package frc.robot.util;
+package frc.robot.util.math;
 
 import edu.wpi.first.math.MathSharedStore;
