+++ src/main/java/frc/robot/util/swerve/DrivingMotor.java
@@ -1,4 +1,4 @@
-package frc.robot.util;
+package frc.robot.util.swerve;
 
 public enum DrivingMotor {
     NEO,
