+++ src/main/java/frc/robot/util/swerve/TurningMotor.java
@@ -1,4 +1,4 @@
-package frc.robot.util;
+package frc.robot.util.swerve;
 
 public enum TurningMotor {
     NEO_550,
