+++ src/main/java/frc/robot/util/upper_assembly/ScoringHeight.java
@@ -1,4 +1,4 @@
-package frc.robot.util;
+package frc.robot.util.upper_assembly;
 
 public enum ScoringHeight {
     L1,
