+++ src/main/java/frc/robot/util/upper_assembly/narwhal/NarwhalClimberState.java
@@ -1,4 +1,4 @@
-package frc.robot.util;
+package frc.robot.util.upper_assembly.narwhal;
 
 /**
  * <h2> NarwhalClimberState </h2>
