+++ src/main/java/frc/robot/util/upper_assembly/squid/SquidClimberPosition.java
@@ -1,4 +1,4 @@
-package frc.robot.util;
+package frc.robot.util.upper_assembly.squid;
 
 import frc.robot.Constants.SquidConstants.SquidClimberConstants;
