+++ src/main/java/frc/robot/util/vision/ScoringLocation.java
@@ -1,4 +1,4 @@
-package frc.utils.vision;
+package frc.robot.util.vision;
 
 import edu.wpi.first.math.geometry.Transform3d;
