--- src/main/java/frc/utils/DrivingMotor.java
@@ -1,7 +1,0 @@
-package frc.utils;
-
-public enum DrivingMotor {
-    NEO,
-    KRAKEN_X60,
-    KRAKEN_X60_FOC,
-}
