--- src/main/java/frc/utils/UpperSubstructure.java
@@ -1,7 +1,0 @@
-package frc.utils;
-
-public enum UpperSubstructure {
-    SQUID,
-    NARWAL,
-    NONE,
-}
